requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
ffmpeg-python==0.2.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
lameenc==1.8.1
macholib==1.16.3
Markdown==3.8
modulegraph==0.19.6
numpy==2.2.6
openai==1.80.0
packaging==25.0
pillow==11.2.1
psutil==7.0.0
py2app==0.28.8
PyAudio==0.2.14
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
pynput==1.8.1
pyobjc-core==11.0
pyobjc-framework-ApplicationServices==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-CoreAudio==11.0
pyobjc-framework-CoreText==11.0
pyobjc-framework-Quartz==11.0
PyYAML==6.0.2
requests==2.32.3
rumps==0.4.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
